requirements.txt
fastapi
jinja2
aiofiles
python-multipart 
python-dotenv
requests
replicate
uvicorn
openai
FastAPI
Transformers
Torch
